[Maintenance] refactor to use new maintenance task from pimcore 5.8
