+++ src/CoreShop/Bundle/CoreBundle/Maintenance/ProposalExpireTask.php
@@ -10,14 +10,13 @@
  * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
  */
 
-namespace CoreShop\Bundle\CoreBundle\EventListener\Order\Expire;
+namespace CoreShop\Bundle\CoreBundle\Maintenance;
 
 use CoreShop\Bundle\OrderBundle\Expiration\ProposalExpirationInterface;
 use CoreShop\Component\Core\Configuration\ConfigurationServiceInterface;
-use Pimcore\Event\System\MaintenanceEvent;
-use Pimcore\Model\Schedule\Maintenance\Job;
+use Pimcore\Maintenance\TaskInterface;
 
-final class ProposalExpireMaintenanceManagerListener
+final class ProposalExpireTask implements TaskInterface
 {
     /**
      * @var ConfigurationServiceInterface
@@ -47,11 +46,17 @@ final class ProposalExpireMaintenanceManagerListener
     /**
      * @param ConfigurationServiceInterface $configurationService
      * @param ProposalExpirationInterface   $proposalExpiration
+     * @param string                        $type
      * @param int                           $days
      * @param array                         $params
      */
-    public function __construct(ConfigurationServiceInterface $configurationService, ProposalExpirationInterface $proposalExpiration, $type, $days = 0, $params = [])
-    {
+    public function __construct(
+        ConfigurationServiceInterface $configurationService,
+        ProposalExpirationInterface $proposalExpiration,
+        string $type,
+        int $days = 0,
+        array $params = []
+    ) {
         $this->configurationService = $configurationService;
         $this->proposalExpiration = $proposalExpiration;
         $this->type = $type;
@@ -59,14 +64,11 @@ final class ProposalExpireMaintenanceManagerListener
         $this->params = $params;
     }
 
-    /**
-     * @param MaintenanceEvent $maintenanceEvent
-     */
-    public function registerExpire(MaintenanceEvent $maintenanceEvent)
+    public function execute()
     {
         $lastMaintenance = $this->configurationService->get(sprintf('system.%s.expire.last_run', $this->type));
 
-        if (is_null($lastMaintenance)) {
+        if (null === $lastMaintenance) {
             $lastMaintenance = time() - 90000; //t-25h
         }
 
@@ -74,9 +76,7 @@ final class ProposalExpireMaintenanceManagerListener
 
         //since maintenance runs every 5 minutes, we need to check if the last update was 24 hours ago
         if ($timeDiff > 24 * 60 * 60) {
-            $manager = $maintenanceEvent->getManager();
-
-            $manager->registerJob(new Job(sprintf('coreshop.%s.expire', $this->type), [$this->proposalExpiration, 'expire'], [$this->days, $this->params]));
+            $this->proposalExpiration->expire($this->days, $this->params);
 
             $this->configurationService->set(sprintf('system.%s.expire.last_run', $this->type), time());
         }
