+++ src/CoreShop/Bundle/CoreBundle/Resources/config/services/order.yml
@@ -52,8 +52,8 @@ services:
         arguments:
             - '@coreshop.core.order_shipment.transformer.order_item_to_shipment_item.inner'
 
-    coreshop.cart.expire.maintenance_listener:
-        class: CoreShop\Bundle\CoreBundle\EventListener\Order\Expire\ProposalExpireMaintenanceManagerListener
+    coreshop.maintenance.cart.expire_task:
+        class: CoreShop\Bundle\CoreBundle\Maintenance\ProposalExpireTask
         arguments:
             - '@coreshop.configuration.service'
             - '@coreshop.cart.expire'
@@ -61,10 +61,10 @@ services:
             - '%coreshop.cart.expiration.days%'
             - { anonymous: '%coreshop.cart.expiration.anonymous%', customer: '%coreshop.cart.expiration.customer%'}
         tags:
-            - { name: kernel.event_listener, event: pimcore.system.maintenance, method: registerExpire }
+            - { name: pimcore.maintenance.task, type: coreshop_expire_carts }
 
-    coreshop.order.expire.maintenance_listener:
-        class: CoreShop\Bundle\CoreBundle\EventListener\Order\Expire\ProposalExpireMaintenanceManagerListener
+    coreshop.maintenance.order.expire_task:
+        class: CoreShop\Bundle\CoreBundle\Maintenance\ProposalExpireTask
         arguments:
             - '@coreshop.configuration.service'
             - '@coreshop.order.expire'
@@ -72,7 +72,7 @@ services:
             - '%coreshop.order.expiration.days%'
             - []
         tags:
-            - { name: kernel.event_listener, event: pimcore.system.maintenance, method: registerExpire }
+            - { name: pimcore.maintenance.task, type: coreshop_expire_orders }
 
     coreshop.order.number_generator.prefix_suffix:
         class: CoreShop\Component\Core\Order\NumberGenerator\SaleNumberGenerator
